Novations:
- call schedule - see the general call schedule (schedule for each lesson with
the time, when it starts (ex.: 1 lesson starts at 8:00) and when it ends
(ex.: 1 lesson ends at 9:20)). Right now user can see call schedule only
in the settings, but later there'll be separate page for schedules like this;
- password resetting - if user forgot the password, he now can freely reset it
in the authorization block - he only needs to enter his email, by which he
logs in, and press the 'reset password' button. After that if user with the
specified email exists, then he will get on his email his new password, by
which he can log on further. In future releases there will be necessary
password reset confirmation (something like 'follow the link below, if it's
you, who requested for password reset');
- profile has been renamed to settings page, because the only thing you
can do now in the profile (except for the call schedule) right now and
probably in the future is to work with your settings;
- notifications - see your notifications, when they came, the notification
text, mark them as read or even delete them, and come by the link, which
the notification provided. There's also been added notification settings
block to the settings page, but it's still empty, and will be empty probably
up to v0.5 or further;
- code improvements - updates now will come faster because of some
not-large-but-not-small code improvements, mind of which was bothering
since almost the beginning of the project.

Merge branch 'develop' into main for the v0.3

+++ website/util/actions/settings.php
@@ -4,7 +4,7 @@ require_once $_SERVER['DOCUMENT_ROOT'] . "/util/loader.php";
 require_once $_SERVER['DOCUMENT_ROOT'] . "/util/logging.php";
 require_once $_SERVER['DOCUMENT_ROOT'] . "/util/auth_check.php";
 
-if (isset($_POST['action']) && !is_null($currentUser)) {
+if (isset($_POST['action'], $currentUser)) {
     $action = $_POST['action'];
     $response = [];
 
@@ -29,7 +29,7 @@ if (isset($_POST['action']) && !is_null($currentUser)) {
 
     echo json_encode($response);
 } else {
-    header("Location: /pages/news.php");
+    header("Location: /");
 }
 
 function changePassword($currentUser): array
@@ -46,7 +46,7 @@ function changePassword($currentUser): array
 
             if (password_verify($oldPassword, $currentUser->getPassword())) {
                 $currentUser->setPassword($newPassword);
-                $updateResult = StorageRepository::updateUser($currentUser);
+                $updateResult = UserRepository::updateUser($currentUser);
                 if ($updateResult === true) {
                     $result['status'] = 'success';
                     $result['message_places'] = ['general'];
@@ -104,15 +104,19 @@ function changeAvatar($currentUser): array
 
                     $oldAvatarPath = $currentUser->getAvatarPath();
                     $currentUser->setAvatarPath($newFileName);
-                    $updateResult = StorageRepository::updateUser($currentUser);
+                    $updateResult = UserRepository::updateUser($currentUser);
                     if ($updateResult === true) {
                         $avatarsFolderPath = $_SERVER['DOCUMENT_ROOT'] . '/media/user_avatars/';
-                        unlink($avatarsFolderPath . $oldAvatarPath);
+
+                        if ($oldAvatarPath !== 'default.png')
+                            unlink($avatarsFolderPath . $oldAvatarPath);
                         move_uploaded_file($fileLocation, $avatarsFolderPath . $newFileName);
+
                         $result['status'] = 'success';
                         $result['message'] = 'Аватар успішно змінено.';
                     } else {
                         $currentUser->setAvatarPath($oldAvatarPath);
+
                         $result['status'] = 'failure';
                         $result['message'] = 'Виникла помилка при зміні аватару';
                     }
